Refactor optimization code
