+++ DirectXMesh/DirectXMesh_Desktop_2017_Win10.vcxproj
@@ -389,6 +389,8 @@
     <ClCompile Include="DirectXMeshGSAdjacency.cpp" />
     <ClCompile Include="DirectXMeshNormals.cpp" />
     <ClCompile Include="DirectXMeshOptimize.cpp" />
+    <ClCompile Include="DirectXMeshOptimizeLRU.cpp" />
+    <ClCompile Include="DirectXMeshOptimizeTVC.cpp" />
     <ClCompile Include="DirectXMeshRemap.cpp" />
     <ClCompile Include="DirectXMeshTangentFrame.cpp" />
     <ClCompile Include="DirectXMeshUtil.cpp">
