+++ DirectXMesh/DirectXMesh_Desktop_2017_Win10.vcxproj.filters
@@ -53,6 +53,12 @@
     <ClCompile Include="DirectXMeshVBWriter.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="DirectXMeshOptimizeLRU.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="DirectXMeshOptimizeTVC.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <CLInclude Include="DirectXMeshP.h">
