+++ DirectXMesh/DirectXMesh_WindowsPhone81.vcxproj.filters
@@ -53,6 +53,12 @@
     <ClCompile Include="DirectXMeshVBWriter.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="DirectXMeshOptimizeLRU.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="DirectXMeshOptimizeTVC.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="DirectXMesh.inl">
