+++ DirectXMesh/DirectXMesh_XboxOneXDK_2017.vcxproj.filters
@@ -60,5 +60,11 @@
     <ClCompile Include="DirectXMeshVBWriter.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="DirectXMeshOptimizeLRU.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="DirectXMeshOptimizeTVC.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
